requirements.txt
pandas
biopython
feedparser
google-generativeai
openpyxl
